+++ src/Templates/Reset.php
@@ -26,10 +26,10 @@ class Reset extends WPForms_Template {
 		$this->modal       = '';
 		$this->core        = true;
 		$this->data        = [
-			'field_id' => '2',
+			'field_id' => '3',
 			'fields'   => [
-				'0' => [
-					'id'       => '0',
+				'1' => [
+					'id'       => '1',
 					'type'     => 'text',
 					'label'    => esc_html__( 'Username or Email', 'wpforms-user-registration' ),
 					'required' => '1',
@@ -39,8 +39,8 @@ class Reset extends WPForms_Template {
 						'delete'   => false,
 					],
 				],
-				'1' => [
-					'id'           => '1',
+				'2' => [
+					'id'           => '2',
 					'type'         => 'password',
 					'label'        => esc_html__( 'Password', 'wpforms-user-registration' ),
 					'required'     => '1',
@@ -55,7 +55,7 @@ class Reset extends WPForms_Template {
 			'settings' => [
 				'notification_enable' => '0',
 				'disable_entries'     => '1',
-				'antispam'            => '1',
+				'antispam_v3'         => '1',
 				'ajax_submit'         => '1',
 				'user_reset_hide'     => '1',
 			],